mempolicy/mbind: update syscall tests for weighted interleave
Add the MPOL_WEIGHTED_INTERLEAVE policy to the syscall test suite for
get_mempolicy, set_mempolicy, and mbind.

The functional memory distribution test only tests that the default
behavior of WEIGHTED_INTERLEAVE is the same as INTERLEAVE, since the
values for each node default to '1', and because behavior actually
depends on what node the test runs on.  This test at least demonstrates
the interleave mechanism is in use and works by default as intended.

MPOL_WEIGHTED_INTERLEAVE is ifdef defined because it is not upstream
in libnuma yet, so this ensures compilation.

Signed-off-by: Gregory Price <gregory.price@memverge.com>